update module version mongo,helper

+++ project/models/movie.go
@@ -9,7 +9,7 @@ type Movie struct {
 	mongo.BaseModel
 
 	Title       string    `bson:"title" json:"title,omitempty"`
-	Rate        float32   `bson:"rate" json:"rate,omitempty"`
+	Rate        float64   `bson:"rate" json:"rate,omitempty"`
 	ReleaseDate time.Time `bson:"release_date" json:"release_date,omitempty"`
 	Synopsis    string    `bson:"synopsis" json:"synopsis,omitempty"`
 }
